Lots of minor markup adjustments to keep things clean.  Sought & adjusted markup that can make LaTeX2HTML drop a space in the output.
Added explanation of where to find the installed header files for Unix
& Windows.

Added a paragraph at the top of the high-level functions section
explaining the 'start' parameter, telling what the useful values are
and pointing to the specific explanations of each at the end of the
section.

Added a short section to the "Exception Handling" chapter describing
the deprecation of string exceptions and removal of the -X option in
future releases of the interpreter.

Filled in a few missing descriptions throughout the document, and
filled out a bunch more that were incomplete; many descriptions now
sport slick new exception information.

Re-structured the descripition of the API for complex numbers, clearly
separating the C-type from the Python-type portions.

Commented out the "Debugging" chapter since it hasn't been written.