Fix the problems with inherited "special" attributes (attributes described by a data descriptor, such as __class__ and __dict__).
In typeobject.c:

- _PyType_Lookup() is smarter: when used on a static type, it does
  just one dictionary lookup.

- type_getattr() does the right thing, which involves a variation on
  PyObject_GenericGetAttr(), rather than calling it.

In object.c:

- PyObject_Generic{Get,Set}Attr() now calls _PyType_Lookup() right at
  the start.  For static types, this is the same as the single
  dictionary lookup it did before.  For dynamic types, this correctly
  gives "special" attributes priority over what might be in the
  instance dictionary, even if the "special" attribute is inherited
  from several layers deep.

I was worried about the speed of all this.  But not to worry: a
primitive benchmark suggests that for dynamic type instances a normal
attribute lookup is about as fast as that of a classic class.  Static
type instances are even a bit faster!