Fix the subtype test (the previous checkins broke test_descr.py because the subtype test was being applied in a multiple-inheritance situation).
First, _PyObject_TypeCheck() is renamed to PyType_IsSubtype() and
moved from object.c to typeobject.c, where it belings.

Next, its implementation is changed to use the MRO tuple if it exists;

for bootstrap reasons, it walks the tp_base chain (single inheritance
only) as a fallback.

Finally, issubtype() is deleted and calls to it are replaced with a
call to PyType_IsSubtype().