After the merge, this test was failing on a test that insisted that
  class C(object):
      __slots__ = {}

be an error.

Well, because of the way it was written (slots was converted to a
tuple using PySequence_Tuple()) and because of enhanced behavior of
dictionaries (a dictionary in a sequence context behaves like a list
of its keys) this is no longer an error.

Rather than trying to make this an error, I propose to accept this
defeat, and use a different non-sequence type in the test.

If the __slots__ variable is a dictionary, the values are ignored.
But they might be used by some future tool, e.g. as doc strings, or
type specs, or whatever.