Merge trunk changes into 2.2a2 release branch.
I've discovered another reason not to *require* the new module: new is
not allowed in restricted mode.  Without this patch, this means that
any import from __future__ would fail, and in turn e.g. "import types"
would fail.

I'm beginning to wonder if sticking the CO_ constants in __future__
was such a good idea...