Broke dependence of _num2date and _date2num on datetime objects, and renamed to _ord2ymd and _ymd2ord.  Proper testing of these functions requires exercising negative ordinals and years, and using datetime objects blocked that.
Added a datetime.fromordinal() constructor, and a toordinal() method.
"Something like that" is needed for people who want to build other
calendar systems on top of this basic one.

Changed test_ordinal_conversions() to do exhaustive testing on both
sides of the origin, and every 7th year across the datetime range.
This is less testing than it was doing before, but the test runs much
faster now than can be accounted for by that alone.  I suspect it's
avoiding object hair that makes the difference (the test uses _ord2ymd
and _ymd2ord directly now, never constructing a datetime object).