Implement analysis of scope information for PySymtable_Analyze(). The results are accessible from PyST_GetScope().
A comment explains the details of the analysis; it seems much cleaner
than the old implementation in compile.c, which tried to do everything
in a single pass.

The analysis is incomplete because it (probably) does not handle the
case where a free variable is bound in a scope that declares the name
global.