Gave datetime a correct timetuple implementation.  The tests didn't catch that it was still using the date timetuple() (which zeroes out hours, minutes, and seconds), so beefed up the test of that too.
Changed date.strftime() to call the correct timetuple() method for the
object passed to it.  Added a new test to ensure that strftime is seeing
all the fields it should see for a datetime.strftime().

Updated the datetime docs, to near completion.  There are still a couple
methods that aren't right.