Experimental support for pickle protocol 2.  Classes timedelta, date, time, and datetime now have a __new__ method instead of __init__. (Class tzinfo not yet; I'll have to think more about it.)  I've added __getnewargs__ as an alias for __getstate__ for each of these classes. The state for date is now a tuple (for the others it already was a tuple).
The test suite now tests protocols 0, 1 and 2; this is the only change
(except one place where the state of a date was compared to a string
literal).  It passes for Python 2.2 as well as for Python 2.3.

XXX The pickles are not compatible with those of the C implementation!
XXX But I think the C implementation can be changed to match, and it
XXX will work for all three protocols, in both Python versions.
XXX Fingers crossed.  Goodnight.