Field descriptors for c_char arrays get a setfunc and getfunc as well, this allows to treat them as strings as well.
In _CData_get and CData_set, prefer to use getfunc and setfunc over
the type even if it is not NULL, this allows the above treatment.

In _CData_set, when a field is initialized with a tuple, call the
constructor with the tuple and use the result. Works very nice!

New tests for this. Still have to implement this for c_wchar.