Yet another attempt to make exiting Python robust after having used ctypes.com:
A function is registered with atexit that removes any com objects from
sys.exc_info() - by raising and ignoring an exception.

Then, a global object is created which will call CoUnitialize() when
the ctypes.com module is cleaned up.  It would be possible to delay
this further by sticking it into the __builtin__ module.