Improved error handling inside ast.c; fixes test_syntax.
Add ast_error() that captures line number of error and
ast_finish_error() to assign filename and location.  It's a little
complicated this way, but less intrusive than passing the filename to
every ast function that might raise an exception.

Need to pass filename to PyAST_FromNode() for error handling.