Fix handling of testlist_gexp.
Created a single function that handles the several possible testlist
variants, with a flag indicating whether a gexp is possible.  Note
that I don't understand why that Grammar allows testlist_gexp in some
places but not others.

Address several other XXX comments, including calls to free.  Reindent
a few functions to four spaces so that they can be read.