Add Gregory K. Johnson's revised version of mailbox.py (funded by the 2005 Summer of Code).
The revision adds a number of new mailbox classes that support adding
and removing messages; these classes also support mailbox locking and
default to using email.Message instead of rfc822.Message.

The old mailbox classes are largely left alone for backward compatibility.
The exception is the Maildir class, which was present in the old module 
and now inherits from the new classes.  The Maildir class's interface 
is pretty simple, though, so I think it'll be compatible with existing 
code.

(The change to the NEWS file also adds a missing word to a different
news item, which unfortunately required rewrapping the line.)