Merging of the Stackless core changes (the difference between 2.4.3 release source code and the Stackless 2.4.3 dev source code) into the branched 2.5b1 trunk source code.
It compiles.  The interpreter starts.  But any attempt at running either the Stackless unit tests or the Python regression tests dies horribly with a "SystemError: error return without exception set" (as should be expected).  There are a lot of changes from 2.4 to 2.5 which affect Stackless, and no attempt has been made as yet to do anything bar get the source code to compile and the interpreter to run.  Notable known issues are the introduction of PyEval_EvalFrameEx, which adds a new argument to the evaluation of frames (something core to Stackless) and the new generator functions which coroutine like functionality, which have been left in place unlooked at and are guaranteed to work less than anything else.