Cause a call to load_module() that cannot load a module to raise ImportError (PEP 302 does not specify what to do).
Also fleshed out tests to make sure the proper number of arguments are expected
for find_module() (can be a common error when using static and classmethods;
thanks Paul Moore for the tests).

Finally, comments about the included frozen modules in the stdlib were added so
as to not forget them.