Deleted comment: XXX Reference URLs that compare the different UI packages They are too changeable, so instead I've added a reference to the Python Wiki page which has such links.
Dropped coverage of PyKDE because it is platform-specific. I think
it is best to only list cross-platform GUI libraries here and leave
people to go to the Wiki for platform-specific GUI libraries.

Dropped the listing for FXPy since its website hasn't been updated
since 2002 and its most "up-to-date" installer for Windows is for Python 2.2!

Slightly slimmed down wxWidgets coverage for balance.

So now only the "big 3" cross-platform toolkits,
PyGTK, PyQt, and wxPython, are listed, and they now appear in
alphabetical order.