Flesh out (but with no tests) the source loader.
This led to two other changes.  One was not directly exposes the source and
extension module loaders to the outside world.  This is because they are
heavily dependent on their respective importers.

Second, the arguments passed to the loaders by FileImporter has been expanded.
Now all information gathered by FileImporter.find_module is passed as an
argument.