Last commit here before merging back to the py3k branch head.  This changes the default Header maximum line length to 78 for consistency with RFC 2822 and other interfaces in this package.
Normalize the chunks before generating the output of Header.__str__().

Headers must compare equality against their unicode unencoded string value not
their RFC 2047 encoded value.  This is because the unicode representation is
the canonical one and the encoded representation is influenced by external
factors such as maximum line lengths.

Work in a 'transitional space' for dealing with the RFC 2047 required space
between encoded words on the same line.  This isn't quite right because it
leaves a bogus trailing space on the last line of a header if it ends with an
encoded word.  I'll fix this after the merge back to the py3k branch.

base64mime.body_encode(): Remove the 'binary' argument so I can also kill
fix_eols().