Started documentation for Combobox;
Updated samples to work both in py3k and python 2.x;

Added a Ttk version for py3k, it is under Lib/tkinter/ttk.py (according 
to the stdlib-sig);

Updated README to match where I've tested the ttk module and added one 
more reference;

Checking for empty value in _format_optdict just doing 'if not value' 
was a big mistake, it caused segmentation fault in python 2.5.2 and 
wasn't working correctly with all the options it may be receive. Of 
course it shouldn't be done like that, since values like 0 are accepted, 
it is re-corrected now.