Merged in r62123-64233 from the mainline trunk.
'configure' is still an older version generated from an older version of 'configure.in'.  It still needs to be regenerated and checked in.

There are still deprecation warnings happening when you subclass the tasklet class and have it call the base '__init__' function with the passed in function as an argument when it gets instantiated.  This appears when running the Stackless unittests, in 'test_watchdog.py'.

The mainline 'test_sys.py' unittest is failing because it checks the size of the typeobject according to a formula, which does not take into account what Stackless adds to this structure.

Two other mainline unittests fail, 'test_site.py' and 'test_distutils.py', however, these also fail when Stackless is compiled as normal Python with the Stackless functionality turned off - so are not a Stackless problem.