Remove much of the #ifdef stuff in Gay's code.
In detail
---------
Assume: that we're using IEEE 754 floating-point with the usual
round-half-to-even rounding, that divisions are accurate, that our
float.h contains everything that C89 says it should.

We don't want: K&R headers, Hex nans, Hex floats,
correct setting of the inexact flag, rounding to respect FPU
rounding mode.

We *do* want:  errno to be set appropriately, checks
for infinities and nans.