Merged r64818 from py3k.
When a main tasklet is destroyed, it was expected that there wouldn't be a current tasklet if the recursion count was higher than 0.  This is not correct in the case the tasklet was created by an interpreter command, as each creates a new main tasklet, and the creation of a new tasklet within that will leave it in the scheduler as the current tasklet, as the main exits.

Why did this not error in older versions of Python?  I do not know.  But the fix makes sense, and the recursion count being greater than 0 is valid (it was that way before we created the main tasklet).